Create Forked Fag Off
supposed to be a copy of the ruby  plus the rainbow sades in 50 ways  + all kids of more crap. minus tavf h ohinkingt asw i. he rubying the with and the withour version of the in all possible rainblown color fart graph in zero line space to create a linear interpretation of the math described un the gpt instruction project comand.  so feel free to ask questions and ill answer them as i am putting other aspects of this all encompassing project of mine in the GOD's Plan in my God's Equation  fot the creation of ALLCreation <{[(-*_)]}>=(^-*Transpose) , SQUARE(^*_*"")